+++ ark/docs/src/components/shiki.config.js
@@ -31,18 +31,21 @@ export const twoslash = transformerTwoslash({
 		filterNode: node => {
 			switch (node.type) {
 				case "hover":
+					if (node.text.endsWith(", {}>"))
+						// omit default scope param from type display
+						node.text = node.text.slice(0, -5) + ">"
+					if (node.text.startsWith("type")) {
+						return true
+					}
 					if (node.text.startsWith("const")) {
-						if (node.text.endsWith(", {}>"))
-							// omit default scope param from type display
-							node.text = node.text.slice(0, -5) + ">"
 						// show type with completions populated for known examples
 						node.text = node.text.replace(
-							"isAdmin?: never",
-							"isAdmin?: boolean | null"
+							"version?: never",
+							`version?: number | string`
 						)
 						node.text = node.text.replace(
-							"luckyNumbers: never",
-							"luckyNumbers: (number | bigint)[]"
+							"versions?: never",
+							"versions?: (number | string)[]"
 						)
 						// filter out the type of Type's invocation
 						// as opposed to the Type itself
@@ -50,24 +53,12 @@ export const twoslash = transformerTwoslash({
 					}
 					if (node.text.startsWith(twoslashPropertyPrefix)) {
 						const expression = node.text.slice(twoslashPropertyPrefix.length)
-						if (expression.startsWith("ArkErrors.summary")) {
-							// cleanup runtime errors for display
-							const runtimeErrorSummary = /^ArkErrors\.summary: "(.*)"/.exec(
-								expression
-							)
-							if (runtimeErrorSummary) {
-								node.text = runtimeErrorSummary[1].split("\\n").join("\n")
-							}
+						return (
+							// this shows error summary in JSDoc
+							expression.startsWith("RuntimeErrors.summary") ||
 							// this helps demonstrate narrowing on discrimination
-							return true
-						}
-						if (expression === "luckyNumbers: (number | bigint)[]")
-							// this helps demonstrate narrowing on discrimination
-							return true
-						if (expression.endsWith("typeof ArkErrors"))
-							// also helps clarify how discrimination works
-							return true
-						return false
+							expression === `platform: "android" | "ios"`
+						)
 					}
 					return false
 				case "error":