+++ ark/docs/src/content/docs/intrinsicOptimization.twoslash.js
@@ -1,8 +1,8 @@
 import { type } from "arktype"
 // prettier-ignore
 // ---cut---
-// all unions are optimally discriminated-
-// even at nested paths or in multiple passes!
+// all unions are optimally discriminated
+// even if multiple/nested paths are needed
 const account = type({
 	kind: "'admin'",
 	"powers?": "string[]"