+++ ark/type/__tests__/bounds.test.ts
@@ -2,7 +2,7 @@ import { attest, contextualize } from "@arktype/attest"
 import {
 	internalKeywords,
 	keywordNodes,
-	rawRoot,
+	rawSchema,
 	writeInvalidOperandMessage,
 	writeUnboundableMessage
 } from "@arktype/schema"
@@ -36,7 +36,7 @@ contextualize(
 			const t = type("number<10")
 			attest<number>(t.infer)
 			attest(t).type.toString.snap("Type<lessThan<10>, {}>")
-			const expected = rawRoot({
+			const expected = rawSchema({
 				domain: "number",
 				max: { rule: 10, exclusive: true }
 			})
@@ -47,7 +47,7 @@ contextualize(
 			const t = type("number<=-49")
 			attest<number>(t.infer)
 			attest(t).type.toString.snap("Type<atMost<-49>, {}>")
-			const expected = rawRoot({
+			const expected = rawSchema({
 				domain: "number",
 				max: { rule: -49, exclusive: false }
 			})
@@ -58,15 +58,24 @@ contextualize(
 			const t = type("number==3211993")
 			attest<3211993>(t.infer)
 			attest(t).type.toString.snap("Type<3211993, {}>")
-			const expected = rawRoot({ unit: 3211993 })
+			const expected = rawSchema({ unit: 3211993 })
 			attest(t.json).equals(expected.json)
+		})
+
+		it("== length", () => {
+			const verifyCodeSchema = type({ code: "string==6" })
+
+			attest(verifyCodeSchema({ code: "123456" })).snap({ code: "123456" })
+			attest(verifyCodeSchema({ code: "foo" }).toString()).snap(
+				"code must be exactly length 6 (was 3)"
+			)
 		})
 
 		it("<,<=", () => {
 			const t = type("-5<number<=5")
 			attest(t).type.toString.snap("Type<is<MoreThan<-5> & AtMost<5>>, {}>")
 			attest<number>(t.infer)
-			const expected = rawRoot({
+			const expected = rawSchema({
 				domain: "number",
 				min: { rule: -5, exclusive: true },
 				max: 5
@@ -80,7 +89,7 @@ contextualize(
 				"Type<is<AtLeast<-3.23> & LessThan<4.654>>, {}>"
 			)
 			attest<number>(t.infer)
-			const expected = rawRoot({
+			const expected = rawSchema({
 				domain: "number",
 				min: { rule: -3.23 },
 				max: { rule: 4.654, exclusive: true }
@@ -92,7 +101,7 @@ contextualize(
 			const t = type("number > 3")
 			attest(t).type.toString.snap("Type<moreThan<3>, {}>")
 			attest<number>(t.infer)
-			const expected = rawRoot({
+			const expected = rawSchema({
 				domain: "number",
 				min: { rule: 3, exclusive: true }
 			})